Llama_Q_A_chatbot/requirements.txt
ollama
langchain-community
python-dotenv
streamlit
bs4
requests 
beautifulsoup4
transformers
langchain-chroma
langchain-core
ipykernel